Pig Game
The game is aimed at 2 people. The goal is to roll the dice and hold scores.

You can roll the dice as many times as you wish (all the scores are added) until you get 1. In this case your scores reset to zero and the move is passed to the second player.

The one who first has 20 scores will be the winner.